requirements.txt
garminconnect>=0.2.0
python-telegram-bot>=20.0
APScheduler>=3.10.0
pytz>=2023.3
